Migrated SubModule test to TS
